+++ CHANGELOG.md
@@ -5,7 +5,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 
 ## [Unreleased]
-(PA-3613) Added vanagon support for MacOS 11 Big Sur
+- (PA-3613) Added vanagon support for MacOS 11 Big Sur
+
+- (VANAGON-85) Added the `clear_provisioning` method to the platform dsl to clear 
+the provisioning command array. 
+
 ## [0.21.0] - released 2021-04-15
 ### Added
 - (VANAGON-85) Moved platform definitions into core vanagon. Created default