+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog adheres to [Keep a CHANGELOG](http://keepachangelog.com/).
 ### Added
 - (PA-3709) Add Debian 11 64-bit support
 - (PA-3755) add `--extended-attributes` on mac OS
+- (RE-14154) Exit if `VANAGON_FORCE_SIGNING` is set
 
 ## [0.21.1] - released 2021-06-07
 ### Added